==> notebooks/0.5-SHL-data-prep-creating-a-balanced-set.ipynb <==
import pandas as pd
import os

# set root
os.chdir('/Users/davidharar/Documents/School/thesis/downstream_classification')


# load data

metadata = pd.read_csv('./data/combined_data/metadata_only_existant_readings_09042023.csv',index_col=0)
metadata_train = metadata[metadata['sample'] == 'train']
metadata_not_train = metadata[metadata['sample'] != 'train']

metadata_train_0 = metadata_train[metadata_train['one_year_until_death'] == 0].reset_index(drop=True)
metadata_train_1 = metadata_train[metadata_train['one_year_until_death'] == 1].reset_index(drop=True)

N0 = len(metadata_train_0)
N1 = len(metadata_train_1)

N = max(N0,N1)

metadata_train_1_oversampled = metadata_train_1.sample(n = N, replace=True)
metadata_train_1.shape, '-->', metadata_train_1_oversampled.shape

metadata_train_balanced = pd.concat([metadata_train_0, metadata_train_1_oversampled], axis=0, ignore_index=True)
print("oversampled DM distribution")
display(metadata_train_balanced['one_year_until_death'].value_counts())

metadata_for_testing_balanced_by_DM = pd.concat([metadata_train_balanced, metadata_not_train], axis=0, ignore_index=True)

display(
    pd.crosstab(
        metadata_for_testing_balanced_by_DM['sample'],
        metadata_for_testing_balanced_by_DM['one_year_until_death']
    )
)

metadata_for_testing_balanced_by_DM.sample(frac=1).reset_index(drop=True).to_csv('./data/combined_data/metadata_balanced_by_death.csv')
