==> notebooks/1.5-modeling-build-skeleton-for-5leads-and-lead-extraction-for-analysis.ipynb <==
# Summary

import pandas as pd
import numpy as np
import os

# make sure we're in the correct directory
os.chdir('/Users/davidharar/Documents/School/thesis/downstream_classification')

from dataloader.DataLoader import *

# dataloader
train_generator = DataGenerator(
    metadata_file_path= './data/combined_data/metadata_only_existant_readings_09042023.csv',   # path to metadata file
    data_folder_path = './data/individual-signals/',     # path to individual signals
    sample='train',               # sample we want to create a generator to. Either train, validation or test
    targets=['AF'],             # list of targets we want train on
    batch_size=4,           # batch size
    shuffle=True       # Whether to shuffle the list of IDs at the end of each epoch.
                 )

X,y = train_generator.__getitem__(1224)
print(X.shape, y.shape)
X,y





pd.DataFrame(X[0])

min_number_of_leads = 5
mask = np.array([np.count_nonzero(~np.isnan(X[j][:,0]))>min_number_of_leads for j in range(len(X))])
print(mask)
X[mask]



X[mask].shape, y[mask].shape


# Tweaks to training
- aucpr
- learning rate decay
- patients
- understand - how to use weights


def PRAUC(y_true, y_pred):
    """
    Get area under the precision recall curve.
    """
    from sklearn.metrics import precision_recall_curve, auc
    precision, recall, thresholds = precision_recall_curve(y_true, y_pred)
    return auc(recall, precision)


PRAUC([0,  1,  0,  1,  1], [0.2,0.3,0.7,0.9,0.4])
