==> notebooks/4.0-adding-invalid-leads-create-combined-data.ipynb <==
import os
import numpy as np
import pandas as pd

root = '/home/david/Desktop/projects/thesis/downstream_classification/'
origin = root + 'data/combined_data_all_readings_with_invalid/'
destination = root + 'data/individual-signals-with-invalid/'

if not os.path.exists(destination):
    os.makedirs(destination)

os.chdir(root)



os.listdir(origin)

# Merge

leads = ['LI', 'LII', 'LIII', 'aVF', 'aVL', 'aVR','V1','V2','V3','V4','V5','V6']
def preprocessing(df, lead):
    df['index'] = df['index'].str.replace(f'{lead}_','')
    df.columns = ['index']+[f'{lead}_{j}' for j in range(450)]
    return df



# init by load LI:
combined = pd.read_csv(origin+'LI_all_readings_2909.csv', index_col = 0)
combined  = preprocessing(df = combined, lead='LI')

# merge with more leads
for lead in leads[1:]:
    temp_lead_df = pd.read_csv(origin+f'{lead}_all_readings_2909.csv', index_col = 0)
    temp_lead_df = preprocessing(df = temp_lead_df, lead=lead)

    assert len(temp_lead_df['index']) == len(temp_lead_df), "There are duplicated indices"

    combined = combined.merge(temp_lead_df,
                 on ='index',
                 how = 'outer')
    
print('Total Number of Obs:', len(combined))
combined = combined.dropna()
print('Number of Obs w/out empties:', len(combined))

combined.iloc[:5]

leads[1:]
