==> notebooks/10.3.ptb-data-get-scores-for-imputation-without-training.ipynb <==
# Summary
We want to have the scores under the imputation scenario but without joint tratining.
- Take the code file for joint training
- Modify it to return answers before joint trainng



import os
import sys

import numpy as np
import pickle 


# import trainer
# -----------------------------
os.chdir('/home/david/Desktop/projects/thesis/')
sys.path.append('/home/david/Desktop/projects/thesis/')

from combined_downstream_upstream.utils.LoadModels import *
from combined_downstream_upstream.executors.train_combined_model import trainer, combined_results_without_joint_training_on_test


# upstream model params and config
# ---------------
upstream_params = {
    'input_dimension': 12,
    'output_dimension': 12,              
    'hidden_dimmension':  128,           # d_model (int) – the number of expected features in the input (required)???,
    'attention_heads': 8,               # number of attention heads, if None then d_model//64,
    'encoder_number_of_layers': 8,
    'dropout': 0.4,
    'clip': 1,
    'positional_encodings': False,
    'device':'cuda'
}

upstream_model   = load_upstream_model(upstream_params, folder_path = './upstream_seq2seq/models/', model_name = 'transformer_cnn_4heads')



results = {'turned_off_channels':[], 'roc_auc':[], 'pr_auc':[]}

for channels in range(12):
    try:
        downstream_models_root = f"./downstream_classification/models/AF-V10-different-seeds-ptb-{channels}_channels_off-saving-downstream-models"
        downstream_models_names = [x for x in os.listdir(downstream_models_root) if x.endswith('.pt')]

        for model_path in downstream_models_names:
            downstream_model = load_downstream_model(
                dropout=0.5,scale=1,num_inputs=12, 
                weights_path=f"{downstream_models_root}/{model_path}"
                )

            aucpr, rocauc = combined_results_without_joint_training_on_test(
                metadata_file_path      = './downstream_classification/data/combined_data/ptb_signal_level_metadata_with_label.csv',
                data_folder_path        = './downstream_classification/data/ptb-ecg-processed-divided-into-450/', 
                targets                 = ['AF'],
                batch_size              = 128, 
                impute_only_missing     = True,
                upstream_model          = upstream_model,
                downstream_model        = downstream_model,
                continue_training_upstream_model = False,
                leads = ['LI', 'LII', 'LIII', 'aVL', 'aVR','aVF', 'V1','V2','V3','V4','V5','V6'],
                loss_function_weight = None,

                reconstruction_loss_weight = 0,

                internal_data=False,
                channels_to_turn_off = channels,

                )
            
            results['turned_off_channels'].append(channels)
            results['roc_auc'].append(rocauc)
            results['pr_auc'].append(aucpr)
    except:
        continue

import pandas as pd
results_df = pd.DataFrame(results)
results_df.to_csv("/home/david/Desktop/projects/thesis/downstream_classification/notebooks/imputation_withought_joint_training.csv", index=False)

results_df

roc_auc_se   = pd.DataFrame((results_df.groupby(['turned_off_channels'])['roc_auc'].sem() * 100).round(2)).reset_index()
roc_auc_mean = pd.DataFrame((results_df.groupby(['turned_off_channels'])['roc_auc'].mean() * 100).round(2)).reset_index()
pr_auc_se    = pd.DataFrame((results_df.groupby(['turned_off_channels'])['pr_auc'].sem() * 100).round(2)).reset_index()
pr_auc_mean  = pd.DataFrame((results_df.groupby(['turned_off_channels'])['pr_auc'].mean() * 100).round(2)).reset_index()

roc_auc_df = roc_auc_mean.merge(roc_auc_se, on = 'turned_off_channels')
roc_auc_df['Resulls'] = [f'{x} ({y})' for x,y in zip(roc_auc_df['roc_auc_x'],roc_auc_df['roc_auc_y'])]

pr_auc_df = pr_auc_mean.merge(pr_auc_se, on = 'turned_off_channels')
pr_auc_df['Resulls'] = [f'{x} ({y})' for x,y in zip(pr_auc_df['pr_auc_x'],pr_auc_df['pr_auc_y'])]


display(roc_auc_df,pr_auc_df)

results_df.groupby(['turned_off_channels'])['roc_auc'].mean(),results_df.groupby(['turned_off_channels'])['pr_auc'].mean()
