==> notebooks/12.0-create_imputed_downstream_shl_data.ipynb <==
# Summary
In some of the experiments we run, we impute signals, freeze the upstream weights and basically only running the downstream weights. In these cases, we can just once, impute the downstream data and run it. 


import pandas as pd
import pickle
import matplotlib.pyplot as plt
import seaborn as sns

# combined_model_train_only_downstream = pd.read_csv("/Users/davidharar/Documents/School/thesis/Writing/ptb-upstream-results.csv")
# display(combined_model_train_only_downstream)

# combined_model_train_both = pd.read_csv('/Users/davidharar/Documents/School/thesis/Writing/ptb-upstream-results-unfreeze_upstream_weights.csv')
# display(combined_model_train_both)

results_downstream = []
for channels in range(12):
    
    with open(f'/home/david/Desktop/projects/thesis/combined_downstream_upstream/models/ptb-{channels}_channels_off-no_additional_noise-impute_only_missing-cold_start/best_rocauc_and_pr_auc.pkl', 'rb') as handle:
        b = pickle.load(handle)
    bdf = pd.DataFrame(b)
    results_downstream.append(bdf)

results_downstream = pd.concat(results_downstream,axis=0,ignore_index=True)
results_downstream

# cresults_downstream.to_csv('/home/david/Desktop/projects/thesis/combined_downstream_upstream/notebooks/ptb-imputation-train_only_downstream-cold_start.csv',index=False)
